requirements.txt
customtkinter==5.2.2
pynput==1.7.7
mss==9.0.2
Pillow==10.3.0
pystray==0.19.5
bcrypt==4.1.3
pywin32==306
mysql-connector-python==8.3.0
pyinstaller==6.6.0
